Complete rewrite of parser (now called lexer).
Changed examples/str.hc to have a U64 pack 8 chars into it.
Readme now mentions when cloning the project to use `git clone --recursive`.
Updated the name of src/parser.* -> src/lexer.*. This is the correct name for it's functionality.
Rewrite of src/lexer.c to be less confusing and cleaner. Elements of the previous algorithm reside but it should be less LOC and easier to read.