Conversion done.
Update to `example/bad_constant.hc`, added semicolon, shows use of an invalid base.
Now the `validate_numerical_constant` function works, will convert strings to floats and integers handling errors.
Give the compiler the default _DEFAULT_SOURCE, this will allow the use of strtoll which is not posix, but libc has it
Rewrote the decipher function, was getting strange segfaults and the old version was a bit messy.

Signed-off-by: Alistair <alistair@alistairbell.xyz>